+++ inst/include/realPSD/Model_Template.hpp
@@ -1,16 +1,17 @@
 /// @file {{{Model}}}_Generics.hpp
 /// @brief Generic code which supplies everything needed to the **TMB** compiler for `{{{Model}}}`.
 
-{{#Hide}}
+{{^Standalone}}
 #ifndef {{{Model}}}_Generics_hpp
 #define {{{Model}}}_Generics_hpp 1
-{{/Hide}}
+{{/Standalone}}
 
-{{#Show}}
+{{#Standalone}}
 #include <TMB.hpp>
-{{/Show}}
+{{/Standalone}}
+#include "realPSD/FitMethods.hpp"
 #include "{{{Header}}}" // model class definition
-#include "{{{Include}}}" // the generic package code
+// #include "{{{Include}}}" // the generic package code
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR obj
@@ -23,13 +24,19 @@ Type {{{Model}}}_Generics(objective_function<Type>* obj) {
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR this
 
-{{#Show}}
+{{#Standalone}}
 template<class Type>
 Type objective_function<Type>::operator() () {
-  return {{{Model}}}_Generics<Type>(this);
+  DATA_STRING(model);
+  if(model == "{{{Model}}}") {
+    return {{{Model}}}_Generics<Type>(this);
+  } else {
+    error("Unknown model.");
+  }
+  return Type(0.0);
 }
-{{/Show}}
+{{/Standalone}}
 
-{{#Hide}}
+{{^Standalone}}
 #endif
-{{/Hide}}
+{{/Standalone}}